requirements.txt
ansible
black
flake8
jsonref
invoke
pytest
jsonschema
ruamel.yaml
tox
yamllint
